feat: Store user email in contest participation
This commit adds the user's email to the `contest_participants` table when a user participates in a contest. This is a necessary step to enable future features that may require a user's email for contest-related communications or prize distribution.

The translation for "Winners of the last contest" has been removed as it's no longer directly displayed in the carousel. The admin view for contest details has also been updated to include participant emails and to improve the presentation of winner information.